Move ForgottenDomain struct to data
This type is not really a model. I doesn't have CRUD operations, and it
doesn't act as a representation of a database table.

Types like this (see data.DestinationsCount for another example) are
nothing more than return values for a single data function. Moving it
to the data package makes this more clear. It also makes the code much
easier to trace because you don't have to jump into a completely
separate file and package when looking up the type.