Remove some unused errors
NotImplemented was not used at all.

InternalServerError doesn't need to exist. It's not handled in any
special way, and it's already the default behaviour for an unhandled
error.

Conflict was only used in one place. We have data.UniqueConstraintError
for any conflict in the database, so we can use that.