Remove validate.String function
Functions with more than 3 arguments are hard to use, especially when
many of the arguments are the same type.

Generally the solution to more args is to create a struct as the single
argument. In this case we already have a struct, so we can simply remove
the function.

Also resolve a TODO by removing the use of validate and using regular
switch/case statement.

Also move some constants to the once place they are used.