Use SKIP LOCKED in update queries
Using where updated_at was not sufficient to prevent the row from being
locked.  To avoid the locks use a sub-select with FOR UPDATE SKIP
LOCKED. When this returns no rows (because the row was locked) the
entire update is skipped. This allows us to restore a transaction
around the middleware without blocking other requests from the same
access key.