Fix ListGrants blocking requests with versioned handler
The TestConnector_Run test was flaking pretty often. The problem was
that the API is pinned to v0.18.1, and that version handler had a bug in
it.

We were not setting the response header because the SetHeader function
was on the ListGrantsResponse (not ListResponse), and because we were
not setting the update index value on ListResponse.

This commit fixes both problems, and the test no longer flakes.