Move PolymorphicID constructor functions to models
This change was done automatically with GoLand refactoring tools.

This change is being made in preparation for removing PolymorphicID, and
replacing it with the new Subject type.

By moving this constructor we'll be able to swap the implementation for
one that returns Subject, and we won't need to touch these lines again.