Decode JSON body directly
It turns out the gin binding for JSON is nothing but a wrapper around
the stdlib. We can use the stdlib directly, which removes 3 layers of
indirection.

Also remove indiration in the other two binding calls by calling the
appropriate function directly. This is being done in preparation for
replacing the library that decodes the HTTP request into our application
specific request struct.

Also check if there are values to decode before calling bind. The
binding operation is relatively expensive because it uses reflection to
iterate over the request struct. Checking for inputs prevents the
reflection and iteration.